move PresenceDot file and update styles
