+++ src/app/members/memberCard.tsx
@@ -5,7 +5,7 @@ import Link from "next/link"
 import { Card, CardFooter, Image } from "@nextui-org/react"
 import LikeButton from "@/components/LikeButton"
 import type { Member } from "@prisma/client"
-import PresenceDot from "./PresenceDot"
+import PresenceDot from "../../components/PresenceDot"
 
 type memberCardProps = {
   member: Member
@@ -27,10 +27,10 @@ export default function MemberCard({ member, likeIds }: memberCardProps) {
         width={300}
         className="aspect-square object-cover"
       />
-      <div className="absolute top-[10px] left-[10px] z-10">
+      <div className="absolute top-2 left-2 z-10">
         <PresenceDot member={member} />
       </div>
-      <div className="absolute top-[10px] right-[10px] z-10">
+      <div className="absolute top-3 right-3 z-10">
         <LikeButton isLiked={isLiked} targetUserId={member.userId} />
       </div>
       <CardFooter className="absolute bottom-0 bg-black z-10 bg-dark-gradient">
