+++ src/components/PresenceDot.tsx
@@ -10,15 +10,16 @@ type Props = {
 export default function PresenceDot({ member }: Props) {
   const onlineMembers = usePresenceStore(state => state.members)
   const isOnline = onlineMembers.indexOf(member.userId) !== -1
+
+  if (!isOnline) return null
+
   return (
     <div className="relative">
       <GoDot size={36} className="fill-white" />
-      {isOnline && (
-        <GoDotFill
-          size={32}
-          className="absolute top-[2px] left-[2px] fill-green-500"
-        />
-      )}
+      <GoDotFill
+        size={32}
+        className="absolute top-[2px] left-[2px] fill-green-500"
+      />
     </div>
   )
 }
